webidl2, improvements
